Moved blobIndex into Sofast
