+++ opencsp/app/sofast/lib/ProcessSofastFixed.py
@@ -11,7 +11,7 @@ from opencsp.app.sofast.lib.MeasurementSofastFixed import MeasurementSofastFixed
 from opencsp.app.sofast.lib.ParamsSofastFixed import ParamsSofastFixed
 from opencsp.common.lib.camera.Camera import Camera
 from opencsp.common.lib.csp.MirrorPoint import MirrorPoint
-from opencsp.common.lib.deflectometry.BlobIndex import BlobIndex
+from opencsp.app.sofast.lib.BlobIndex import BlobIndex
 import opencsp.common.lib.deflectometry.calculation_data_classes as cdc
 from opencsp.common.lib.deflectometry.FacetData import FacetData
 import opencsp.app.sofast.lib.image_processing as ip
